apps/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
faiss-cpu==1.8.0
numpy==1.26.4
sentence-transformers==3.0.1
requests==2.32.3
pydantic==2.8.2
